requirements.txt

azure-functions
requests
python-dotenv
dlt
duckdb
dbt-core
duckdb-extensions
duckdb-extension-motherduck

